refactor(config): split `menu.ts` into modular files
