+++ extensions/omarchy-menu/src/config/about.ts
@@ -1,0 +1,9 @@
+import { Icon } from "@vicinae/api";
+import { MenuItem } from "./types";
+
+export const about: MenuItem = {
+  id: "about",
+  name: "About",
+  icon: Icon.Info,
+  command: "omarchy-launch-about",
+};
