+++ extensions/omarchy-menu/src/config/types.ts
@@ -1,0 +1,7 @@
+export type MenuItem = {
+  id: string;
+  name: string;
+  icon: string;
+  command?: string;
+  items?: MenuItem[];
+};
